requirements.txt
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
flask==3.0.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
functions-framework==3.5.0
gunicorn==21.2.0 